Move `ZzzScene.cpp/h` to `Scenes` with new name SceneCore for improved project structure consistency
