+++ src/source/CSQuest.cpp
@@ -20,7 +20,7 @@
 #include "ZzzObject.h"
 #include "ZzzTexture.h"
 #include "ZzzCharacter.h"
-#include "ZzzScene.h"
+#include "Scenes/SceneCore.h"
 #include "ZzzInterface.h"
 #include "ZzzInventory.h"
