+++ src/source/CSWaterTerrain.cpp
@@ -16,7 +16,7 @@
 #include "ZzzObject.h"
 #include "ZzzTexture.h"
 #include "ZzzCharacter.h"
-#include "ZzzScene.h"
+#include "Scenes/SceneCore.h"
 #include "ZzzInterface.h"
 #include "ZzzInventory.h"
 #include "ZzzLodTerrain.h"
